Resolve prod index migration issue

+++ db/migrate/20201007013846_add_unique_index_slug_blog_posts.rb
@@ -1,6 +1,6 @@
 class AddUniqueIndexSlugBlogPosts < ActiveRecord::Migration[6.0]
   def change
-    remove_index :blog_posts, :slug
+    remove_index :blog_posts, :slug if index_exists?(:blog_posts, :slug)
     add_index :blog_posts, :slug, unique: true
   end
 end
